HTML

+++ src/main/java/com/examen/DAW_CL1_MARCO_ARANDA/model/bd/Especialidad.java
@@ -22,5 +22,5 @@ import lombok.Setter;
         private String nomesp;
 
         @Column(name="costo")
-        private Integer costo;
+        private Double costo;
 }
